+++ src/app/(provider)/(root)/cart/_components/index.ts
@@ -1,4 +1,4 @@
-export { default as CartDrawer } from './CartDrawer';
+export { default as CartAccordion } from './CartAccordion';
 export { default as CartEmpty } from './CartEmpty';
 export { default as CartItem } from './CartItem';
 export { default as CartList } from './CartList';
